Add arm/arm64 support (#280)
Add arm/arm64 support

Add IElfHeader interface, ElfHeaderCommon, ElfHeader64, ElfHeader32 structs

Add IElfPRStatus interface, ElfPRStatusX64, ElfPRStatusArm, ElfPRStatusArm64 structs

Add ArmContext, Arm64Context and X86Context and made them public because any code that
gets the thread context needs these structs.

Add both 32 and 64 bit versions of the ElfProgramHeaders, ElfSectionHeaders, etc.

Add RegSetArm and RegSetArm64 structs.

+++ src/Microsoft.Diagnostics.Runtime/src/Linux/LinuxLiveDataReader.cs
@@ -220,7 +220,7 @@ namespace Microsoft.Diagnostics.Runtime.Linux
                 return false;
             }
             AMD64Context* ctx = (AMD64Context*)context.ToPointer();
-            ctx->ContextFlags = (int)contextFlags;
+            ctx->ContextFlags = contextFlags;
             IntPtr ptr = Marshal.AllocHGlobal(sizeof(RegSetX64));
             try
             {
@@ -248,7 +248,7 @@ namespace Microsoft.Diagnostics.Runtime.Linux
             }
             IntPtr ptrContext = Marshal.AllocHGlobal(sizeof(AMD64Context));
             AMD64Context* ctx = (AMD64Context*)ptrContext;
-            ctx->ContextFlags = (int)contextFlags;
+            ctx->ContextFlags = contextFlags;
             IntPtr ptr = Marshal.AllocHGlobal(sizeof(RegSetX64));
             try
             {
